Improved printo, added %s, %d and %c (Like printf) on printo. Added itoa.c.

+++ Kernel/libc/memory/memcpy.c
@@ -1,6 +1,6 @@
 #include <types.h>
 
-void *memcpy(void *dest, void *src,unsigned int n){
+void *memcpy(void *dest, void *src,size_t n){
   char* _dest=dest;
   char* _src=src;
   while(n){ /* While n!=0 */
