+++ Kernel/libc/memory/memset.c
@@ -1,4 +1,6 @@
-void *memset(void *dst,int c,unsigned int n){
+#include <types.h>
+
+void *memset(void *dst,int c,size_t n){
   char* _dst=dst;
   while(n--){
     *_dst++=(char)c;
